Allow overriding the mounted filename
The default behavior is to use the file referenced in the HTTP(S) URL
as the mounted filename. This is not always desirable however since the
URL can be very long or contain characters that will cause problems for
the filesystem. Thus, allow specifying a filename via a new command line
option.